# api/openapi/kapsora-v1.yaml
openapi: 3.1.0
info:
  title: KAPSORA Platform API
  version: 1.0.0-alpha.1
  summary: Kurumsal hak, fayda ve hizmet orkestrasyonu için başlangıç API sözleşmesi
  description: |
    Bu sözleşme ilk sprint ve çekirdek MVP için contract-first geliştirme başlangıç noktasıdır.
    Kullanıcılar KAPSORA'nın kendi hesaplarıyla giriş yapar (ADR-022); tarayıcı yalnız opak
    oturum çerezi taşır. Makineden makineye entegrasyonlar OAuth 2.0 client credentials ve
    mTLS ile çalışır. Tüm tenant kapsamlı çağrılar X-Tenant-ID
    başlığı taşır. Durum değiştiren POST komutlarında Idempotency-Key zorunludur; sorgu
    niteliğindeki POST'larda (eligibility check, pricing quote) opsiyoneldir.
    İstemci X-Request-ID başlığında UUID gönderebilir; sunucu her yanıtta X-Request-ID döner.
    Çerez oturumuyla yapılan durum değiştiren çağrılar X-CSRF-Token başlığı taşır.
  contact:
    name: KAPSORA Product Engineering
servers:
  - url: https://api.staging.kapsora.example
    description: Staging
  - url: https://api.kapsora.example
    description: Production
security:
  - sessionCookie: []
  - oauth2: [kapsora.api]
...
paths:
...
  /api/v1/contracts/{contractId}/versions:
    get:
      tags: [Contracts]
      operationId: listContractVersions
      description: |
        The version summaries of one contract, highest version number first. The price
        sheet of a version is read through the version itself; this list is what a screen
        shows to pick between "what we agreed then" and "what we agreed now".
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractId'
      responses:
        '200':
          description: Contract versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersionList'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Contracts]
      operationId: createContractVersion
      description: |
        Opens the next DRAFT version of a contract. copyFromVersionId copies the price
        lists, price items, package definitions, quotas and payment term of another
        version of the same contract, which is how a yearly price revision starts from
        last year's sheet instead of from an empty page. The new version stays editable
        until it is published.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContractVersionRequest'
      responses:
        '201':
          description: Contract version created
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/contract-versions/{contractVersionId}:
    get:
      tags: [Contracts]
      operationId: getContractVersion
      description: |
        One contract version with its price lists, its review metadata and its
        configuration hash. The price items of each list are paged separately, because a
        health tariff runs to thousands of rows. A published version needs only
        `contract.read`; a draft needs `contract.manage`, so a price nobody has agreed to
        yet is not visible to everyone who may read the agreed one.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
      responses:
        '200':
          description: Contract version
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Contracts]
      operationId: patchContractVersion
      description: |
        Merge-patch of the validity period, the currency and the notes of a DRAFT version.
        A version that is not DRAFT answers 409 CONTRACT_VERSION_IMMUTABLE: everything
        downstream was priced from the published sheet, so it cannot move underneath a
        claim that already quoted it.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateContractVersionRequest'
      responses:
        '200':
          description: Contract version updated
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '415':
          description: Body is not application/merge-patch+json
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/contract-versions/{contractVersionId}/submit:
    post:
      tags: [Contracts]
      operationId: submitContractVersion
      description: |
        Moves a DRAFT version to UNDER_REVIEW and records the maker. A version with no
        price list holding at least one price item is refused with 422 on field
        priceLists: an empty sheet cannot be reviewed, and publishing it would make every
        later price lookup answer PRICE_NOT_FOUND. The validity start is required too,
        because a published version without one could never be selected by service date.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewComment'
      responses:
        '200':
          description: Contract version submitted for review
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/contract-versions/{contractVersionId}/publish:
    post:
      tags: [Contracts]
      operationId: publishContractVersion
      description: |
        Freezes a version under review. It needs `contract.publish`, a recent step-up and
        an actor other than the one who submitted it; the same person answers 403
        MAKER_CHECKER_SAME_ACTOR and the refusal is audited. Publishing writes a
        configuration hash over the version's whole price content, so "what was agreed"
        can be proved later. Two published versions of one contract may not cover the same
        date: an overlap answers 409 CONTRACT_VERSION_OVERLAP.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewComment'
      responses:
        '200':
          description: Contract version published
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/contract-versions/{contractVersionId}/retire:
    post:
      tags: [Contracts]
      operationId: retireContractVersion
      description: |
        Closes a published version with a reason. It needs `contract.publish` and a recent
        step-up. The version stays readable for every quote and claim priced from it and
        stops being a candidate for new price lookups; the reason code is required so the
        history says why the tariff was withdrawn.
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/ContractVersionId'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonCommand'
      responses:
        '200':
          description: Contract version retired
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractVersion'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          description: ETag mismatch
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          description: If-Match header missing
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-kapsora_session
      description: |
        Opaque session id. The server stores only its SHA-256 hash and the browser never
        holds a token of any kind. Local HTTP development uses the name kapsora_session
        because the __Host- prefix requires HTTPS.
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://identity.kapsora.example/realms/kapsora/protocol/openid-connect/token
          scopes:
            kapsora.api: Call KAPSORA APIs
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: Selected tenant UUID. It must be one of the actor's active memberships.
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated unique key retained for at least 24 hours.
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required when the request is authenticated with the BFF session cookie.
      schema:
        type: string
        minLength: 16
        maxLength: 256
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Optimistic concurrency token returned as ETag.
      schema:
        type: string
...
    ContractId:
      name: contractId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ContractVersionId:
      name: contractVersionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    ETag:
      description: Strong optimistic concurrency tag based on row_version.
      schema:
        type: string
  responses:
...
    Forbidden:
      description: Authenticated actor lacks permission or scope
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource does not exist or is not visible in the selected tenant
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Business state, duplicate idempotency key, or uniqueness conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationError:
      description: Request is syntactically valid but violates validation or business preconditions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_.-]{2,79}$'
        traceId:
          type: string
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, code]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
...
    ReviewComment:
      type: object
      additionalProperties: false
      properties:
        comment:
          type: string
          maxLength: 1000
...
    ReasonCommand:
      type: object
      additionalProperties: false
      required: [reasonCode]
      properties:
        reasonCode:
          type: string
          minLength: 2
          maxLength: 80
        reasonText:
          type: string
          maxLength: 1000
...
    ContractVersionStatus:
      type: string
      description: |
        Lifecycle of a contract version. PUBLISHED is the only status a quote, an
        authorization or a claim ever reads, and nothing about it may change afterwards.
      enum: [DRAFT, UNDER_REVIEW, PUBLISHED, RETIRED]
...
    ContractVersionSummary:
      type: object
      additionalProperties: false
      required: [id, contractId, versionNo, status, currencyCode, rowVersion]
      properties:
        id:
          type: string
          format: uuid
        contractId:
          type: string
          format: uuid
        versionNo:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/ContractVersionStatus'
        validFrom:
          type: [string, 'null']
          format: date
        validTo:
          type: [string, 'null']
          format: date
        currencyCode:
          type: string
          pattern: '^[A-Z]{3}$'
        notes:
          type: [string, 'null']
        publishedAt:
          type: [string, 'null']
          format: date-time
        rowVersion:
          type: integer
          minimum: 1
    ContractVersionList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ContractVersionSummary'
    ContractVersion:
      type: object
      additionalProperties: false
      required: [id, contractId, versionNo, status, currencyCode, rowVersion, priceLists]
      properties:
        id:
          type: string
          format: uuid
        contractId:
          type: string
          format: uuid
        versionNo:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/ContractVersionStatus'
        validFrom:
          type: [string, 'null']
          format: date
        validTo:
          type: [string, 'null']
          format: date
        currencyCode:
          type: string
          pattern: '^[A-Z]{3}$'
        notes:
          type: [string, 'null']
        configurationHash:
          type: [string, 'null']
          description: Lower-case hex SHA-256 over the version's whole price content, written at publish.
        submittedAt:
          type: [string, 'null']
          format: date-time
        submittedBy:
          type: [string, 'null']
          format: uuid
        publishedAt:
          type: [string, 'null']
          format: date-time
        publishedBy:
          type: [string, 'null']
          format: uuid
        reviewComment:
          type: [string, 'null']
        retireReasonCode:
          type: [string, 'null']
        priceLists:
          type: array
          items:
            $ref: '#/components/schemas/PriceList'
        rowVersion:
          type: integer
          minimum: 1
    CreateContractVersionRequest:
      type: object
      additionalProperties: false
      properties:
        copyFromVersionId:
          type: string
          format: uuid
          description: Another version of the same contract whose whole price content is copied.
        validFrom:
          type: string
          format: date
        validTo:
          type: [string, 'null']
          format: date
        currencyCode:
          type: string
          pattern: '^[A-Z]{3}$'
        notes:
          type: string
          maxLength: 2000
    UpdateContractVersionRequest:
      type: object
      additionalProperties: false
      description: Merge-patch body of a DRAFT contract version.
      properties:
        validFrom:
          type: [string, 'null']
          format: date
        validTo:
          type: [string, 'null']
          format: date
        currencyCode:
          type: string
          pattern: '^[A-Z]{3}$'
        notes:
          type: [string, 'null']
          maxLength: 2000
    PriceList:
      type: object
      additionalProperties: false
      required: [id, contractVersionId, code, name, priority, rowVersion]
      properties:
        id:
          type: string
          format: uuid
        contractVersionId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        priority:
          type: integer
          description: Higher wins when two equally specific price items tie.
        seasonFrom:
          type: [string, 'null']
          format: date
        seasonTo:
          type: [string, 'null']
          format: date
        weekdayMask:
          type: [integer, 'null']
          minimum: 1
          maximum: 127
          description: Bit 0 is Monday through bit 6 Sunday; null means every day.
        itemCount:
          type: integer
          minimum: 0
          description: How many price items the list currently holds.
        rowVersion:
          type: integer
          minimum: 1
...